bridge/requirements.txt
pyserial>=3.5
websockets>=12.0
aiohttp>=3.9.0
